фикс физ*ки

+++ xrODE/include/ode/config.h
@@ -20,182 +20,93 @@
  *                                                                       *
  *************************************************************************/
 
-
 /* per-machine configuration */
 
-
-
 #ifndef _ODE_CONFIG_H_
-
 #define _ODE_CONFIG_H_
 
 #ifdef __cplusplus
-
 extern "C" {
-
 #endif
-
-
 
 #include <stdio.h>
-
 #include <stdarg.h>
-
 #include <malloc.h>		// for alloca under windows
-
 #include <string.h>
-
 #include <math.h>
 
-
-
 #if defined(WIN32) && (defined(MSVC) || defined(MINGW))
-
 static union { unsigned char __c[4]; float __f; } __ode_huge_valf =
-
   {{0,0,0x80,0x7f}};
-
 #define _INFINITY4 (__ode_huge_valf.__f)
-
 static union { unsigned char __c[8]; double __d; } __ode_huge_val =
-
   {{0,0,0,0,0,0,0xf0,0x7f }};
-
 #define _INFINITY8 (__ode_huge_val.__d)
-
 #else
-
 #define _INFINITY8 HUGE_VAL
-
 #define _INFINITY4 HUGE_VALF
-
 #endif
-
-
 
 #if defined(dSINGLE)
-
 #define dInfinity _INFINITY4
-
 #else
-
 #define dInfinity _INFINITY8
-
 #endif
 
-
-
 #define SHAREDLIBIMPORT __declspec (dllimport)
-
 #define SHAREDLIBEXPORT __declspec (dllexport)
 
-
-
 /* some types. assume `int' >= 32 bits */
-
 typedef unsigned int    uint;
-
 typedef int             int32;
-
 typedef unsigned int    uint32;
-
 typedef short           int16;
-
 typedef unsigned short  uint16;
-
 typedef char            int8;
-
 typedef unsigned char   uint8;
 
 
-
-
-
 /* an integer type that we can safely cast a pointer to and from without
-
  * loss of bits.
-
  */
-
-typedef uintptr_t intP;
-
-
-
+typedef uint32 intP;
 
 
 /* if we're compiling on a pentium, we may need to know the clock rate so
-
  * that the timing function can report accurate times. this number only needs
-
  * to be set accurately if we're doing performance tests - otherwise just
-
  * ignore this. i have not worked out how to determine this number
-
  * automatically yet.
-
  */
 
-
-
 #ifdef PENTIUM
-
 #ifndef PENTIUM_HZ
-
 #define PENTIUM_HZ (496.318983e6)
-
 #endif
-
 #endif
-
-
-
 
 
 /* the efficient alignment. most platforms align data structures to some
-
  * number of bytes, but this is not always the most efficient alignment.
-
  * for example, many x86 compilers align to 4 bytes, but on a pentium it is
-
  * important to align doubles to 8 byte boundaries (for speed), and the 4
-
  * floats in a SIMD register to 16 byte boundaries. many other platforms have
-
  * similar behavior. setting a larger alignment can waste a (very) small
-
  * amount of memory. NOTE: this number must be a power of two.
-
  */
-
-
 
 #define EFFICIENT_ALIGNMENT 16
 
 
-
-
-
 /* for unix, define this if your system supports anonymous memory maps
-
  * (linux does).
-
  */
-
-
 
 #define MMAP_ANONYMOUS
 
 
-
-
-
 #ifdef __cplusplus
-
 }
-
 #endif
 
-
-
 #endif
-
